GuestRequest 컴포넌트화
