+++ sublet-front/src/pages/GuestRequest/Components/ConfilrmButton.js
@@ -1,0 +1,13 @@
+import * as s from '../styles/GuestRequest.styles';
+
+
+
+export function ComfilrmButton() {
+	return (
+		<div className="mt-4">
+			<s.confilrmButton onClick={() => RequestPost(inputs)}>
+				확인하기
+			</s.confilrmButton>
+		</div>
+	);
+};
