+++ sublet-front/src/pages/GuestRequest/Components/UserSelectInfo/Info3/index.js
@@ -1,0 +1,12 @@
+
+import { Contract } from "./Contract"
+import { NumberRoom } from "./NumberRoom"
+
+export function Info3() {
+	return (
+		<div>
+			<Contract />
+			<NumberRoom />
+		</div>
+	)
+}
