+++ sublet-front/src/pages/GuestRequest/Components/UserSelectInfo/Info4/index.js
@@ -1,0 +1,11 @@
+import { BuildingType } from "./BuildingType"
+import { LimitPeople } from "./LimitPeople"
+
+export function Info4() {
+    return (
+        <div>
+            <BuildingType />
+            <LimitPeople />
+        </div>
+    )
+}
