+++ sublet-front/src/pages/GuestRequest/Components/UserSelectInfo/Info5/index.js
@@ -1,0 +1,11 @@
+import { NumberBedroom } from "./NumberBedroom"
+import { NumberBathroom } from "./NumberBathroom"
+
+export function Info5() {
+    return (
+        <div>
+            <NumberBedroom />
+            <NumberBathroom />
+        </div>
+    )
+}
