+++ sublet-front/src/pages/GuestRequest/Components/UserSelectInfo/RequestWord/UserRequestWord.js
@@ -1,0 +1,16 @@
+import GuestRequeststyles from "../../styles/GuestRequest.styles";
+
+export function RequestWord() {
+	const styles = GuestRequeststyles();
+
+	return (
+		<div>
+			<p style={styles.GuestRequest_DetailTitle}>요청 사항</p>
+			<textarea
+				id="message"
+				rows="4"
+				className="block p-2.5 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+				placeholder="Write your thoughts here..."></textarea>
+		</div>
+	);
+};
